Implemented V1 Program
Finished the initial version of the program. A finite state machine
is used to keep track of where the program is and what the timer
is being used form. The TIMER0 hardware device is used in conjunction
with the TIMER0 overflow interrupt to keep track of how much time
is spent in the various states.